updated MissingNumber class to allow for 7-digit numbers; added a performance test to MissingNumberTest

+++ src/main/java/dev/ericdrake/Main.java
@@ -2,7 +2,7 @@ package dev.ericdrake;
 
 public class Main {
     public static void main(String[] args) {
-        int missingNumber = MissingNumber.missingNumber(new int[]{2, 1, 4, 3, 6, 5, 7, 10, 9});
+        long missingNumber = MissingNumber.missingNumber(new long[]{2, 1, 4, 3, 6, 5, 7, 10, 9});
         System.out.println(missingNumber);
 
         String evalString = "((this is some text) more text)(even more text))";
