+++ src/main/java/dev/ericdrake/MissingNumber.java
@@ -3,9 +3,9 @@ package dev.ericdrake;
 import java.util.Arrays;
 
 public class MissingNumber {
-    public static int missingNumber(int[] numbers){
+    public static long missingNumber(long[] numbers){
         Arrays.sort(numbers);
-        int previousNumber = numbers[0];
+        long previousNumber = numbers[0];
         //System.out.println(Arrays.toString(numbers));
         for(int i = 1; i < numbers.length; i++){
             //System.out.println("Number: " + numbers[i]);
